LayoutTests/workers/sab/postMessage-clones.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body>
<script>
description("Checks that window.postMessage clones SharedArrayBuffers");

if (window.testRunner)
    testRunner.waitUntilDone();

var sab = new SharedArrayBuffer(4);
var memory = new Int32Array(sab);
var otherMemory;

window.addEventListener("message", function (event) {
    otherMemory = event.data;
    memory[0] = 42;
    shouldBe("memory[0]", "42");
    shouldBe("otherMemory[0]", "0");
    otherMemory[0] = 43;
    shouldBe("memory[0]", "42");
    shouldBe("otherMemory[0]", "43");
    if (window.testRunner)
        testRunner.notifyDone();
});

postMessage(memory, "*");
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
