LayoutTests/workers/sab/postMessage-clones-expected.txt

Checks that window.postMessage clones SharedArrayBuffers

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
PASS memory[0] is 42
PASS otherMemory[0] is 0
PASS memory[0] is 42
PASS otherMemory[0] is 43
